Add files via upload
The MERN Stack Bookstore project is a comprehensive web application designed to manage a bookstore's inventory, facilitate online purchases, and provide an engaging user experience for both customers and administrators. Leveraging the power of the MERN stack, this project combines a robust backend with a dynamic frontend to create a seamless book buying experience.